requirements.txt
streamlit
pandas
numpy
scikit-learn
pyyaml
requests
plotly
